# EnterpriseWarehouse.Wasm/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "EnterpriseWarehouse.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Product": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "Возвращает список всех товаров.",
        "responses": {
          "200": {
            "description": "Список товаров успешно возвращён.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Product"
        ],
        "summary": "Добавляет новый товар.",
        "requestBody": {
          "description": "Объект EnterpriseWarehouse.API.DTO.ProductCreateDTO, содержащий информацию о новом товаре.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Товар успешно добавлен.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/Product/{id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "Возвращает информацию о товаре по его идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор товара.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Товар найден и информация успешно возвращена.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDTO"
                }
              }
            }
          },
          "404": {
            "description": "Товар с указанным идентификатором не найден."
          }
        }
      },
      "put": {
        "tags": [
          "Product"
        ],
        "summary": "Обновляет информацию о существующем товаре.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор товара.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Объект EnterpriseWarehouse.API.DTO.ProductCreateDTO, содержащий обновлённые данные товара.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Товар успешно обновлён.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDTO"
                }
              }
            }
          },
          "404": {
            "description": "Товар с указанным идентификатором не найден."
          }
        }
      },
      "delete": {
        "tags": [
          "Product"
        ],
        "summary": "Удаляет товар по его идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор товара.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Товар успешно удалён."
          },
          "404": {
            "description": "Товар с указанным идентификатором не найден."
          }
        }
      }
    },
...
    "/api/Query/products-recived-on-date": {
      "get": {
        "tags": [
          "Query"
        ],
        "summary": "Получает список продуктов, полученных указанной организацией на заданную дату.",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Название организации.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "in": "query",
            "description": "Дата поставки.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Query/max-supplies-organizations": {
      "get": {
        "tags": [
          "Query"
        ],
        "summary": "Получает список организаций с максимальным количеством поставок за указанный период.",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "description": "Начальная дата периода.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "Конечная дата периода.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganizationMaxSuppliesDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganizationMaxSuppliesDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganizationMaxSuppliesDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OrganizationMaxSuppliesDTO": {
        "required": [
          "address",
          "geometry",
          "id",
          "name",
          "totalQuantity"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Название",
            "nullable": true
          },
          "address": {
            "type": "string",
            "description": "Адрес",
            "nullable": true
          },
          "geometry": {
            "type": "string",
            "description": "Геометрия в виде полигона",
            "nullable": true
          },
          "totalQuantity": {
            "type": "integer",
            "description": "Количество полученной продукции",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "DTO организации"
      },
...
      "ProductCreateDTO": {
        "required": [
          "code",
          "name"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "Код товара",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Название",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO для создания товара"
      },
      "ProductDTO": {
        "required": [
          "code",
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор",
            "format": "int32"
          },
          "code": {
            "type": "integer",
            "description": "Код товара",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Название",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO для товара"
      },
...
    }
  }
}